Revert "[FSTORE-1453] Move client, decorators, variable_api and constants to hopsworks_common (#229)"
This reverts commit 85deafd4da93c99151a0e996ffbcfb66a2893185.
