--- python/hopsworks/client/online_store_rest_client.py
@@ -1,28 +1,0 @@
-#
-#   Copyright 2024 Hopsworks AB
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-#
-
-from hopsworks_common.client.online_store_rest_client import (
-    OnlineStoreRestClientSingleton,
-    get_instance,
-    init_or_reset_online_store_rest_client,
-)
-
-
-__all__ = [
-    OnlineStoreRestClientSingleton,
-    get_instance,
-    init_or_reset_online_store_rest_client,
-]
