+++ python/tests/core/test_online_store_rest_client.py
@@ -13,10 +13,10 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
-import hopsworks_common
+import hsfs
 import pytest
 from furl import furl
-from hopsworks_common.client import auth, exceptions, online_store_rest_client
+from hsfs.client import auth, exceptions, online_store_rest_client
 
 
 class MockExternalClient:
@@ -50,15 +50,13 @@ class TestOnlineStoreRestClient:
         def client_get_instance():
             return MockExternalClient()
 
-        monkeypatch.setattr(
-            hopsworks_common.client, "get_instance", client_get_instance
-        )
+        monkeypatch.setattr(hsfs.client, "get_instance", client_get_instance)
         variable_api_mock = mocker.patch(
-            "hopsworks_common.core.variable_api.VariableApi.get_loadbalancer_external_domain",
+            "hsfs.core.variable_api.VariableApi.get_loadbalancer_external_domain",
             return_value="app.hopsworks.ai",
         )
         ping_rdrs_mock = mocker.patch(
-            "hopsworks_common.client.online_store_rest_client.OnlineStoreRestClientSingleton.is_connected",
+            "hsfs.client.online_store_rest_client.OnlineStoreRestClientSingleton.is_connected",
         )
 
         # Act
@@ -88,16 +86,14 @@ class TestOnlineStoreRestClient:
         def client_get_instance():
             return MockInternalClient()
 
-        monkeypatch.setattr(
-            hopsworks_common.client, "get_instance", client_get_instance
-        )
+        monkeypatch.setattr(hsfs.client, "get_instance", client_get_instance)
         variable_api_mock = mocker.patch(
-            "hopsworks_common.core.variable_api.VariableApi.get_service_discovery_domain",
+            "hsfs.core.variable_api.VariableApi.get_service_discovery_domain",
             return_value="consul",
         )
         optional_config = {"api_key": "provided_api_key"}
         ping_rdrs_mock = mocker.patch(
-            "hopsworks_common.client.online_store_rest_client.OnlineStoreRestClientSingleton.is_connected",
+            "hsfs.client.online_store_rest_client.OnlineStoreRestClientSingleton.is_connected",
         )
 
         # Act
